+++ archived/webapp-comms-scripts/webcam_image_subscriber.py
